% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{AdjustVariance}
\alias{AdjustVariance}
\title{Adjust feature variance to remove effect of varying feature means. Adapted from pagoda2 [insert citation here]}
\usage{
AdjustVariance(counts, gam.k = 5, plot = F, max.adjusted.variance = 1000,
  min.adjusted.variance = 0.001, verbose = T, q.val = 0.05)
}
\arguments{
\item{counts}{Input data matrix}

\item{gam.k}{Number of additive models to use when fitting mean variance relationship}

\item{plot}{Whether or not to plot the mean variance relationship}

\item{max.adjusted.variance}{Maximum adjusted feature variance}

\item{min.adjusted.variance}{Minimum adjusted feature variance}

\item{verbose}{Whether or not to print output}

\item{q.val}{Maximum adjusted p-value for a feature to qualify as overdispersed}
}
\value{
Dataframe with adjusted feature variances
}
\description{
Adjust feature variance to remove effect of varying feature means. Adapted from pagoda2 [insert citation here]
}
