% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{CalcGenesetLoadings}
\alias{CalcGenesetLoadings}
\title{Calculate geneset factor loadings from gene factor loadings}
\usage{
CalcGenesetLoadings(W, W.genesets.project, top.genesets)
}
\arguments{
\item{W}{gene factor loadings from NMF}

\item{W.genesets.project}{gene factor loadings from projecting genesets onto data}

\item{top.genesets}{List of genesets to calculate loadings for}
}
\value{
Matrix of geneset by factor loadings
}
\description{
Calculate geneset factor loadings from gene factor loadings
}
