% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{CalcSNN}
\alias{CalcSNN}
\title{SNN Graph Construction. Adapted from Seurat.}
\usage{
CalcSNN(data.use, k = 10, k.scale = 10, prune.SNN = 1/15)
}
\arguments{
\item{data.use}{Features x samples matrix to use to build the SNN}

\item{k}{Defines k for the k-nearest neighbor algorithm}

\item{k.scale}{Granularity option for k.param}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard distances when
computing the neighborhood overlap for the SNN construction.}
}
\value{
Returns similarity matrix in sparse matrix format
}
\description{
SNN Graph Construction. Adapted from Seurat.
}
