% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{CheckGeneEmbedding}
\alias{CheckGeneEmbedding}
\title{Validates gene embeddings by plotting cluster logFC vs gene factor loading logFC
Warns users if embedded genes fall below minimum logFC threshold}
\usage{
CheckGeneEmbedding(W, norm.counts, genes.embed, sample.groups, n.bins = 50,
  min.logfc = 2, font.size = 12, label.size = 5, eps = 1e-04)
}
\arguments{
\item{W}{Gene loadings matrix}

\item{norm.counts}{Normalized gene expression matrix}

\item{genes.embed}{Genes to embed onto SWNE plot}

\item{sample.groups}{Sample groupings}

\item{n.bins}{Number of bins for the hexbin plot}

\item{min.logfc}{Minimum logFC for an embedded gene}

\item{font.size}{Axis label font size}

\item{label.size}{Gene label size}

\item{eps}{Pseudocount to add to fold-change calculations}
}
\value{
ggplot2 object
}
\description{
Validates gene embeddings by plotting cluster logFC vs gene factor loading logFC
Warns users if embedded genes fall below minimum logFC threshold
}
