% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{clean_genesets}
\alias{clean_genesets}
\title{Helper function for filtering genesets}
\usage{
clean_genesets(genesets, min.size = 5, max.size = 500, annot = FALSE)
}
\description{
Helper function for filtering genesets
}
