% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{ClusterIndicatorMatrix}
\alias{ClusterIndicatorMatrix}
\title{Creates a clusters x cells indicator matrix where a 1 means the cell belongs to that cluster}
\usage{
ClusterIndicatorMatrix(clusters)
}
\arguments{
\item{clusters}{Cell clusters}
}
\value{
clusters x cells indicator matrix
}
\description{
Creates a clusters x cells indicator matrix where a 1 means the cell belongs to that cluster
}
