% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{ComponentAssociation}
\alias{ComponentAssociation}
\title{Find the top gene or geneset markers for each NMF component using pearson/spearman correlation,
or mutual information}
\usage{
ComponentAssociation(component.scores, feature.mat, n.cores = 8,
  metric = "IC")
}
\arguments{
\item{component.scores}{NMF component scores (nmf components x samples)}

\item{feature.mat}{Feature matrix (features x samples)}

\item{n.cores}{Number of cores to use}

\item{metric}{Association metric. Options are pearson, spearman, or IC}
}
\value{
Components x features matrix of associations
}
\description{
Find the top gene or geneset markers for each NMF component using pearson/spearman correlation,
or mutual information
}
