% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paga.R
\name{compute_partitions}
\alias{compute_partitions}
\title{Compute significant links between specified clusters
Adapted from monocle3 (https://github.com/cole-trapnell-lab/monocle3)}
\usage{
compute_partitions(g, optim_res, qval_thresh = 0.05, verbose = FALSE)
}
\description{
Compute significant links between specified clusters
Adapted from monocle3 (https://github.com/cole-trapnell-lab/monocle3)
}
