% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compute_snn}
\alias{compute_snn}
\title{Helper function for calculating a SNN graph
Adapted from Seurat}
\usage{
compute_snn(train.cell.names, k, train.nn.large, train.nn.ranked,
  test.cell.names = NULL, test.nn.large = NULL, test.nn.ranked = NULL,
  prune.SNN = 1/15, print.output = T)
}
\description{
Helper function for calculating a SNN graph
Adapted from Seurat
}
