% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{DesignMatrix}
\alias{DesignMatrix}
\title{Convert groups list to design matrix}
\usage{
DesignMatrix(groups.list, max.groups = 2, min.cells = 5, drop.cells = T)
}
\arguments{
\item{groups.list}{Named list of cell vectors for each genotype}

\item{max.groups}{Maximum number of genotypes per cell}

\item{min.cells}{Minumum number of cells per genotype}

\item{drop.cells}{Drop cells in genotypes that are filtered out}
}
\value{
Sparse binary matrix where rows are cells, and columns are genotypes. 1 means the cell received the genotype.
}
\description{
Convert groups list to design matrix
}
