% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{EmbedContours}
\alias{EmbedContours}
\title{Embeds data for generating elliptical contours to mark areas where features are present}
\usage{
EmbedContours(swne.embedding, contour.feature, sample.coords = NULL,
  levels.use = NULL, cutoff.use = NULL)
}
\arguments{
\item{swne.embedding}{Existing swne embedding from EmbedSWNE}

\item{contour.feature}{Feature to draw confidence ellipse for, either a factor or numeric vector}

\item{sample.coords}{Sample coordinates to use (default is the sample coordinates from the swne embedding)}

\item{cutoff.use}{If contour.feature is numeric, the cutoff to use to binarize the vector for defining the contour}

\item{levels.plot}{If contour.feature is a factor, specify a subset of levels to generate contours for}
}
\value{
swne.embedding with contour coordinates added
}
\description{
Embeds data for generating elliptical contours to mark areas where features are present
}
