% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{EmbedFeatures}
\alias{EmbedFeatures}
\title{Embeds features relative to factor coordinates}
\usage{
EmbedFeatures(swne.embedding, feature.nmf.assoc, alpha.exp = 1,
  n_pull = NULL)
}
\arguments{
\item{swne.embedding}{Existing swne embedding from EmbedSWNE}

\item{alpha.exp}{Increasing alpha.exp increases how much the factors "pull" the features}

\item{n_pull}{Number of factors pulling on each feature. Must be >= 3}
}
\value{
swne.embedding with feature coordinates (feature.coords)
}
\description{
Embeds features relative to factor coordinates
}
