% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin.R
\name{EmbedTFBS}
\alias{EmbedTFBS}
\title{Embeds genes based on TF binding site accessibility relative to topic coordinates}
\usage{
EmbedTFBS(
  swne.embedding,
  cisTopicObject,
  motif_ix_mat = NULL,
  dev.mat = NULL,
  genes.embed,
  peaks.use = NULL,
  alpha.exp = 1,
  n_pull = 3,
  scale.cols = T,
  overwrite = T
)
}
\arguments{
\item{swne.embedding}{SWNE embedding computed using RunSWNE}

\item{cisTopicObject}{cisTopic object}

\item{motif_ix_mat}{A binary matrix of peaks x TF motifs (can be generated with ChromVar)}

\item{dev.mat}{A matrix of TF binding site accessibility generated from ChromVar}

\item{genes.embed}{Genes to embed}

\item{peaks.use}{Subset of peaks to use}

\item{alpha.exp}{Increasing alpha.exp increases how much the factors "pull" the features}

\item{n_pull}{Number of factors pulling on each feature. Must be >= 3}

\item{scale.cols}{Whether or not to scale the input columns to 0 - 1}

\item{overwrite}{Whether or not to overwrite any existing feature embedding}
}
\value{
swne.embedding with feature coordinates (feature.coords)
}
\description{
Embeds genes based on TF binding site accessibility relative to topic coordinates
}
