% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ExtractCCASeurat}
\alias{ExtractCCASeurat}
\title{Embeds features relative to factor coordinates}
\usage{
ExtractCCASeurat(se.obj)
}
\arguments{
\item{se.obj}{Seurat object with multiple batches aligned via CCA alignment}
}
\value{
Reconstructed gene expression matrix from the aligned CCs
}
\description{
Embeds features relative to factor coordinates
}
