% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ExtractDebatchedSeurat}
\alias{ExtractDebatchedSeurat}
\title{Reconstructs the gene expression matrix from the CCA gene loadings and cell embeddings}
\usage{
ExtractDebatchedSeurat(se.obj)
}
\arguments{
\item{se.obj}{Seurat object with multiple batches aligned via CCA alignment}
}
\value{
Reconstructed gene expression matrix
}
\description{
Reconstructs the gene expression matrix from the CCA gene loadings and cell embeddings
}
