% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ExtractNormCounts}
\alias{ExtractNormCounts}
\title{Extracts scaled counts from Seurat or Pagoda2 objects}
\usage{
ExtractNormCounts(
  obj,
  obj.type = "seurat",
  rescale = T,
  rescale.method = "log",
  batch = NULL
)
}
\arguments{
\item{obj}{Seurat or Pagoda2 object}

\item{obj.type}{Must be either "seurat" or "pagoda2"}

\item{rescale}{Rescale data with plain pagoda2 model}

\item{rescale.method}{Rescale method, either "log", or "ft"}

\item{batch}{If rescaling and multiple batches, specify batch}
}
\value{
Normalized gene expression matrix
}
\description{
Extracts scaled counts from Seurat or Pagoda2 objects
}
