% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{ExtractSWNEColors}
\alias{ExtractSWNEColors}
\title{Extracts the exact colors used to plot each cluster (the hex codes) for a given color seed}
\usage{
ExtractSWNEColors(swne.embedding, sample.groups, seed)
}
\arguments{
\item{swne.embedding}{SWNE embedding (list of factor and sample coordinates) from EmbedSWNE}

\item{sample.groups}{Factor defining sample groups}

\item{seed}{Seed for sample groups color reproducibility}
}
\value{
vector of color hex codes with clusters as the name
}
\description{
Extracts the exact colors used to plot each cluster (the hex codes) for a given color seed
}
