% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{FeaturePlotSWNE}
\alias{FeaturePlotSWNE}
\title{Plots swne embedding with feature overlayed}
\usage{
FeaturePlotSWNE(H.coords, sample.coords, feature.scores, n.colors = 5,
  alpha.plot = 0.5, quantiles = c(0.05, 0.95), samples.plot = NULL,
  label.size = 4.5, pt.size = 1)
}
\arguments{
\item{H.coords}{NMF coordinates}

\item{sample.coords}{Sample coordinates}

\item{n.colors}{Number of colors to bin feature vector into}

\item{alpha.plot}{Data point transparency}

\item{quantiles}{Quantiles to trim outliers from}

\item{samples.plot}{Samples to actually plot. Default is NULL, which plots all samples}

\item{label.size}{Label font size}

\item{pt.size}{Sample point size}

\item{feature.score}{Feature vector to overlay}
}
\value{
ggplot2 object with swne plot with feature overlayed
}
\description{
Plots swne embedding with feature overlayed
}
