% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{FilterData}
\alias{FilterData}
\title{Filter data by minimum cells, minimum genes, and winsorizes data}
\usage{
FilterData(
  x,
  min.samples.frac,
  trim,
  min.nonzero.features = 500,
  max.sample.sum = 50000
)
}
\arguments{
\item{x}{Input matrix}

\item{min.samples.frac}{Minimum number of samples for a feature to be included}

\item{trim}{Fraction of samples to winsorize if trim < 1. If trim > 1, absolute number of samples to winsorize}

\item{min.nonzero.features}{Minimum number of nonzero features for a sample to be included}

\item{max.sample.sum}{Samples with greater than this sum of features will be excluded}
}
\value{
Filtered and winsorized data matrix
}
\description{
Filter data by minimum cells, minimum genes, and winsorizes data
}
