% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{FilterGenesets}
\alias{FilterGenesets}
\title{Filters genesets according to minimum and maximum set size}
\usage{
FilterGenesets(genesets, gene.names, min.size = 5, max.size = 500)
}
\arguments{
\item{gene.names}{Name of all genes to use}

\item{min.size}{Minimum geneset size}

\item{max.size}{Maximum geneset size}
}
\value{
Filtered list of genesets
}
\description{
Filters genesets according to minimum and maximum set size
}
