% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{genesets_indicator}
\alias{genesets_indicator}
\title{Creates a genes x genesets indicator matrix}
\usage{
genesets_indicator(genesets, inv = F, return.numeric = F)
}
\description{
Creates a genes x genesets indicator matrix
}
