% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{get_factor_coords}
\alias{get_factor_coords}
\title{Calculates the coordinates of the NMF factors via Sammon mapping}
\usage{
get_factor_coords(H, method = "sammon", distance = "IC")
}
\description{
Calculates the coordinates of the NMF factors via Sammon mapping
}
