% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{get_sample_coords}
\alias{get_sample_coords}
\title{Calculate sample coordinates using the NMF scores and NMF factor coordinates}
\usage{
get_sample_coords(H, H.coords, alpha = 1, n_pull = NULL)
}
\description{
Calculate sample coordinates using the NMF scores and NMF factor coordinates
}
