% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{ggBarplot}
\alias{ggBarplot}
\title{Make simple barplot from vector}
\usage{
ggBarplot(x, y.lim = NULL, fill.color = "lightgrey")
}
\arguments{
\item{x}{Named numeric vector}

\item{y.lim}{Max y-axis}

\item{fill.color}{Bar color}
}
\value{
barplot
}
\description{
Make simple barplot from vector
}
