% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{ggHeat}
\alias{ggHeat}
\title{Plots a heatmap using ggplot2}
\usage{
ggHeat(
  m,
  rescaling = "none",
  clustering = "none",
  labCol = T,
  labRow = T,
  border = F,
  heatscale = c(low = "skyblue", mid = "white", high = "tomato"),
  legend.title = NULL,
  x.lab.size = 10,
  y.lab.size = 10,
  dot.highlight.cutoff = Inf
)
}
\arguments{
\item{m}{Matrix to plot heatmap for}

\item{rescaling}{Scale by row or columns}

\item{clustering}{Cluster by row or columns}

\item{labCol}{Label columns}

\item{labRow}{Label rows}

\item{border}{Plot heatmap border}

\item{heatscale}{Color map}

\item{legend.title}{Title of heatmap legend}

\item{x.lab.size}{x label size}

\item{y.lab.size}{y label size}

\item{dot.highlight.cutoff}{Heatmap cells above this value  will be highlighted with a dot}
}
\value{
ggplot2 heatmap
}
\description{
Plots a heatmap using ggplot2
}
