% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{ica_init}
\alias{ica_init}
\title{Independent component analysis initialization.
Negative values are set to a small, random number.}
\usage{
ica_init(A, k)
}
\description{
Independent component analysis initialization.
Negative values are set to a small, random number.
}
