% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{MutualInf}
\alias{MutualInf}
\title{Compute Information Coefficient [IC]
Pablo Tamayo Dec 30, 2015}
\usage{
MutualInf(x, y, n.grid = 25)
}
\arguments{
\item{x}{Input vector x}

\item{y}{Input vector y}

\item{n.grid}{Gridsize for calculating IC}
}
\value{
Mutual information between x and y
}
\description{
Compute Information Coefficient [IC]
Pablo Tamayo Dec 30, 2015
}
