% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Helper function for normalizing vectors with different methods.}
\usage{
normalize_vector(x, method = "scale", n_ranks = 10000)
}
\description{
Helper function for normalizing vectors with different methods.
}
