% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{NormalizeCounts}
\alias{NormalizeCounts}
\title{Normalization and batch effect removal adapted from pagoda2 [insert citation here]}
\usage{
NormalizeCounts(counts, depthScale = 1000, batch = NULL)
}
\arguments{
\item{counts}{Input data matrix}

\item{depthScale}{Rescale all columns to this value}

\item{batch}{Factor the same length as ncol(counts) which contains batch effects to remove}
}
\value{
Normalized data matrix with batch effects removed
}
\description{
Normalization and batch effect removal adapted from pagoda2 [insert citation here]
}
