% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{PlotDims}
\alias{PlotDims}
\title{Plots 2D embedding}
\usage{
PlotDims(dim.scores, sample.groups = NULL, x.lab = "tsne1",
  y.lab = "tsne2", main.title = NULL, pt.size = 1, font.size = 12,
  alpha.plot = 1, do.label = T, label.size = 4, show.legend = T,
  show.axes = T, seed = NULL, colors.use = NULL)
}
\arguments{
\item{dim.scores}{2D embedding coordinates. Must be N x 2 samples}

\item{sample.groups}{Factor defining sample groups}

\item{x.lab}{X axis label}

\item{y.lab}{Y axis label}

\item{main.title}{Main plot title}

\item{pt.size}{Sample point size}

\item{font.size}{Font size for axis labels}

\item{alpha.plot}{Data point transparency}

\item{do.label}{Label the sample groups}

\item{label.size}{Label font size}

\item{show.legend}{If sample groups defined, show legend}

\item{show.axes}{Plot x and y axes}

\item{seed}{Seed for sample group color reproducibility}

\item{colors.use}{Vector of hex colors for each sample group. Vector names must align with sample.groups}
}
\value{
ggplot2 object with 2d plot
}
\description{
Plots 2D embedding
}
