% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{PlotFactorSelection}
\alias{PlotFactorSelection}
\title{Plot decrease in reconstruction error versus random noise.
The point where the change in reconstruction error matches the change in reconstruction
error for the randomized matrix is the optimal number of factors to use.}
\usage{
PlotFactorSelection(k.err.diff, font.size = 12)
}
\arguments{
\item{k.err.diff}{Reduction in reconstruction error above noise output by FindNumFactors}

\item{font.size}{Font size to use for plotting}
}
\description{
Plot decrease in reconstruction error versus random noise.
The point where the change in reconstruction error matches the change in reconstruction
error for the randomized matrix is the optimal number of factors to use.
}
