% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{PlotSWNE}
\alias{PlotSWNE}
\title{Plots swne embedding}
\usage{
PlotSWNE(swne.embedding, sample.groups, alpha.plot = 0.25,
  do.label = F, label.size = 4.5, pt.size = 1, samples.plot = NULL,
  show.legend = T, seed = NULL, colors.use = NULL,
  use.brewer.pal = F, contour.geom = "path", contour.alpha = 0.25)
}
\arguments{
\item{swne.embedding}{SWNE embedding (list of factor and sample coordinates) from EmbedSWNE}

\item{sample.groups}{Factor defining sample groups}

\item{alpha.plot}{Data point transparency}

\item{do.label}{Label the sample groups}

\item{label.size}{Label font size}

\item{pt.size}{Sample point size}

\item{samples.plot}{Vector of samples to plot. Default is NULL, which plots all samples}

\item{show.legend}{If sample groups defined, show legend}

\item{seed}{Seed for sample groups color reproducibility}

\item{colors.use}{Vector of hex colors for each sample group. Vector names must align with sample.groups}

\item{use.brewer.pal}{Use RColorBrewer 'Paired' palette instead of default ggplot2 color wheel}

\item{contour.geom}{Plot contour as either a line or a filled in region}

\item{contour.alpha}{Transparency of the contour line/region}
}
\value{
ggplot2 object with swne plot
}
\description{
Plots swne embedding
}
