% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{PlotSWNEVelocyto}
\alias{PlotSWNEVelocyto}
\title{Plots swne embedding with RNA velocity arrows overlaid}
\usage{
PlotSWNEVelocyto(swne.embedding, swne.arrows, alpha.plot = 0.25,
  sample.groups = NULL, arrow.lwd = 0.5, arrow.alpha = 0.5,
  head.size = 0.006, do.label = F, label.size = 4.5, pt.size = 1,
  samples.plot = NULL, show.legend = T, seed = NULL)
}
\arguments{
\item{swne.embedding}{SWNE embedding (list of factor and sample coordinates) from EmbedSWNE}

\item{swne.arrows}{DataFrame of Veloctyo arrow info from VeloctyoArrows}

\item{alpha.plot}{Data point transparency}

\item{sample.groups}{Factor defining sample groups}

\item{arrow.lwd}{Arrow width}

\item{arrow.alpha}{Arrow transparency}

\item{head.size}{Arrowhead size}

\item{do.label}{Label the sample groups}

\item{label.size}{Label font size}

\item{pt.size}{Sample point size}

\item{samples.plot}{Vector of samples to plot. Default is NULL, which plots all samples.}

\item{show.legend}{If sample groups defined, show legend}

\item{seed}{Seed for sample groups color reproducibility}
}
\value{
ggplot2 object with swne plot and RNA velocity arrows
}
\description{
Plots swne embedding with RNA velocity arrows overlaid
}
