% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{ProjectFeatures}
\alias{ProjectFeatures}
\title{Projects new features onto existing factor decomposition}
\usage{
ProjectFeatures(newdata, H, alpha = rep(0, 3), loss = "mse",
  n.cores = 1)
}
\arguments{
\item{newdata}{New data matrix}

\item{H}{Existing factor scores}

\item{alpha}{Regularization parameter}

\item{loss}{Loss function to use}

\item{n.cores}{Number of cores to use}
}
\value{
W matrix (feature loadings) for newdata
}
\description{
Projects new features onto existing factor decomposition
}
