% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{ProjectGenesets}
\alias{ProjectGenesets}
\title{Projects dataset onto genesets using either NMF or a nonnegative linear model}
\usage{
ProjectGenesets(norm.counts, genesets, method = "nnlm", loss = "mse")
}
\arguments{
\item{norm.counts}{Normalized input data}

\item{genesets}{List of genesets to use}

\item{method}{Method used to project data onto genesets. nnlm uses a simple linear model while nmf will use a masked version of nmf.}

\item{loss}{Loss function for the nonnegative linear model}
}
\value{
A genesets x samples matrix
}
\description{
Projects dataset onto genesets using either NMF or a nonnegative linear model
}
