% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{ProjectNMF}
\alias{ProjectNMF}
\title{Projects new data onto existing NMF decomposition}
\usage{
ProjectNMF(newdata, W, alpha = rep(0, 3), loss = "mse")
}
\arguments{
\item{newdata}{New data matrix}

\item{W}{Existing feature loadings}

\item{alpha}{Regularization parameter}

\item{loss}{Loss function to use}
}
\value{
H matrix (scores) for newdata
}
\description{
Projects new data onto existing NMF decomposition
}
