% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{ProjectSamples}
\alias{ProjectSamples}
\title{Projects new samples onto existing factor decomposition}
\usage{
ProjectSamples(newdata, W, alpha = rep(0, 3), loss = "mse", n.cores = 1)
}
\arguments{
\item{newdata}{New data matrix}

\item{W}{Existing feature loadings}

\item{alpha}{Regularization parameter}

\item{loss}{Loss function to use}

\item{n.cores}{Number of cores to use}
}
\value{
H matrix (scores) for newdata
}
\description{
Projects new samples onto existing factor decomposition
}
