% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paga.R
\name{PruneSNN}
\alias{PruneSNN}
\title{Prune SNN with PAGA graph. Adapted from Monocle3}
\usage{
PruneSNN(snn, knn, clusters = NULL, qval.cutoff = 0.05)
}
\arguments{
\item{snn}{Shared Nearest Neighbors (SNN) graph in the form of a sparse matrix}

\item{knn}{kNN in the form of a sparse matrix}

\item{clusters}{Cell clusters}

\item{qval.cutoff}{Q-value cutoff for significant edges}
}
\value{
SNN graph with spurious edges pruned
}
\description{
Prune SNN with PAGA graph. Adapted from Monocle3
}
