% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{Read10XHuman}
\alias{Read10XHuman}
\title{Reads a 10X matrix aligned to a mouse/human reference and pulls out only the human cells
Adapted from Seurat}
\usage{
Read10XHuman(matrix.dir, species.class.file)
}
\arguments{
\item{matrix.dir}{The input matrix or sparse matrix directory}

\item{species.class}{Species classification (from Seurat secondary)}
}
\value{
Returns the human cells and genes
}
\description{
Reads a 10X matrix aligned to a mouse/human reference and pulls out only the human cells
Adapted from Seurat
}
