% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ReadData}
\alias{ReadData}
\title{Reads data from a sparse or dense matrix}
\usage{
ReadData(matrix.dir, make.sparse = T)
}
\arguments{
\item{matrix.dir}{The input matrix or sparse matrix directory}

\item{make.sparse}{Force output to be a sparse dgCMatrix}
}
\value{
Returns either a dense or sparse matrix
}
\description{
Reads data from a sparse or dense matrix
}
