% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{ReadGenesets}
\alias{ReadGenesets}
\title{Load genesets from gmt file. For more info on the .gmt file format, see [link]}
\usage{
ReadGenesets(genesets.file)
}
\arguments{
\item{geneset.file}{Geneset filename}
}
\value{
List of genesets (character vectors)
}
\description{
Load genesets from gmt file. For more info on the .gmt file format, see [link]
}
