% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{ReadGroups}
\alias{ReadGroups}
\title{Read in sample groups from a csv file format}
\usage{
ReadGroups(groups.file, sep.char = ",")
}
\arguments{
\item{groups.file}{Name of groups file}

\item{sep.char}{Delimiter}
}
\value{
List of groups
}
\description{
Read in sample groups from a csv file format
}
