% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{RenameNMFs}
\alias{RenameNMFs}
\title{Function for renaming NMF components to something more interpretable.
If the NMF component name is the empty string, "", then the NMF component will not be plotted}
\usage{
RenameNMFs(swne.embedding, new.names)
}
\arguments{
\item{swne.embedding}{List of NMF and sample coordinates from EmbedSWNE}

\item{new.names}{Character vector of new NMF names. Must be of length equal to the number of NMF components}
}
\value{
SWNE embedding with NMFs renamed
}
\description{
Function for renaming NMF components to something more interpretable.
If the NMF component name is the empty string, "", then the NMF component will not be plotted
}
