% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{RunGSEA}
\alias{RunGSEA}
\title{Runs GSEA on the gene association coefficients for each NMF}
\usage{
RunGSEA(gene.factor.assoc, genesets, power = 1, n.rand = 1000,
  n.cores = 1)
}
\arguments{
\item{gene.factor.assoc}{Matrix with the gene associations for each nmf}

\item{genesets}{Genesets to use (as a named list)}

\item{power}{GSEA coefficient power}

\item{n.rand}{Number of permutations to use when calculating significance}

\item{n.cores}{Number of threads to use}
}
\value{
Dataframe summarizing the gsea results
}
\description{
Runs GSEA on the gene association coefficients for each NMF
}
