% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{RunNMF}
\alias{RunNMF}
\title{Runs NMF decomposition on a data matrix}
\usage{
RunNMF(A, k, alpha = 0, init = "random", n.cores = 1, loss = "mse",
  n.rand.init = 5, init.zeros = "uniform")
}
\arguments{
\item{A}{Input data matrix}

\item{k}{Number of NMF components}

\item{alpha}{Regularization parameter}

\item{init}{NMF initialization method}

\item{n.cores}{Number of cores}

\item{loss}{Type of loss function to use}

\item{n.rand.init}{If random initialization is used, number of random restarts}

\item{init.zeros}{What to do with zeros in the initialization}
}
\value{
List of NMF loadings (features x nmf components) and NMF scores (nmf components x samples)
}
\description{
Runs NMF decomposition on a data matrix
}
