% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ScaleCounts}
\alias{ScaleCounts}
\title{Normalize, adjust feature variance, and scale data matrix}
\usage{
ScaleCounts(counts, batch = NULL, method = "log", adj.var = T,
  plot.var.adj = F)
}
\arguments{
\item{counts}{Input data matrix}

\item{batch}{Factor with batch identity for each sample}

\item{method}{Scaling method}

\item{adj.var}{Whether or not to apply mean variance adjustment for features}

\item{plot.var.adj}{Whether or not to plot the mean variance relationship}
}
\value{
Normalized and scaled data matrix
}
\description{
Normalize, adjust feature variance, and scale data matrix
}
