% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{SelectFeatures}
\alias{SelectFeatures}
\title{Select overdispersed features}
\usage{
SelectFeatures(counts, batch = NULL, n.features = 3000, gam.k = 10)
}
\arguments{
\item{counts}{Input data matrix}

\item{n.features}{Number of features to keep}

\item{gam.k}{Number of additive models to use for variance modeling}
}
\description{
Select overdispersed features
}
