% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{SelectGenesets}
\alias{SelectGenesets}
\title{Select genesets for embedding}
\usage{
SelectGenesets(norm.counts, nmf.scores, genesets,
  assoc.metric = "pearson", method = "lm", n.cores = 8,
  genesets.return = 5)
}
\arguments{
\item{norm.counts}{Normalized input data}

\item{nmf.scores}{NMF factors (factors x samples)}

\item{genesets}{List of genesets to use}

\item{method}{Method used to project data onto genesets. lm uses a simple linear model while mf
will use a masked version of nmf.}

\item{n.cores}{Number of cores to use}

\item{genesets.return}{Number of genesets to return for each factor}
}
\value{
dataframe of top genesets associated with each factor
}
\description{
Select genesets for embedding
}
