% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{single_groups}
\alias{single_groups}
\title{Helper function for filtering out cells belonging to more than one group}
\usage{
single_groups(groups.list, min.cells = 1)
}
\description{
Helper function for filtering out cells belonging to more than one group
}
