% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{SummarizeAssocFeatures}
\alias{SummarizeAssocFeatures}
\title{Summarize the components x features association matrix into a readable dataframe}
\usage{
SummarizeAssocFeatures(component.feature.assoc, features.return = 10,
  features.use = NULL)
}
\arguments{
\item{component.feature.assoc}{NMF components x features association matrix from ComponentAssociation}

\item{features.return}{Number of top features to return for each NMF component}

\item{features.use}{Only consider a subset of features. Default is NULL, which will use all features}
}
\value{
Dataframe summarizing top features for each NMF component
}
\description{
Summarize the components x features association matrix into a readable dataframe
}
