% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{UnflattenDataframe}
\alias{UnflattenDataframe}
\title{Convert dataframe to matrix, specifying all column names}
\usage{
UnflattenDataframe(df, output.name, row.col = "Gene",
  col.col = "Group")
}
\arguments{
\item{df}{Dataframe}

\item{output.name}{Column of df to use as matrix values}

\item{row.col}{Column of df to use as matrix rows}

\item{col.col}{Column of df to use as matrix columns}
}
\value{
Matrix
}
\description{
Convert dataframe to matrix, specifying all column names
}
