% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{UnflattenGroups}
\alias{UnflattenGroups}
\title{Converts a flat sample groups character vector into a list format}
\usage{
UnflattenGroups(groups)
}
\arguments{
\item{groups}{Character vector of sample groups}
}
\value{
List of groups: each list element contains the samples for that group
}
\description{
Converts a flat sample groups character vector into a list format
}
