% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{VelocytoArrows}
\alias{VelocytoArrows}
\title{Function for computing arrows to add onto a ggplot2 object.
Adapted from velocyto.R}
\usage{
VelocytoArrows(emb, vel, pca.red = NULL, n = 200, scale = "sqrt",
  corr.sigma = 0.05, show.grid.flow = T, grid.n = 30, grid.sd = NULL,
  min.grid.cell.mass = 0.5, min.arrow.size = NULL, arrow.scale = 1,
  max.grid.arrow.length = NULL, n.cores = 16, diffusion.steps = 10,
  cc = NULL)
}
\arguments{
\item{emb}{2D embedding (SWNE, t-SNE, etc.) coordinates. Must be N x 2 samples}

\item{vel}{RNA velocity object}

\item{pca.red}{PCA reduction. Must be N x nPCs samples.}
}
\description{
Function for computing arrows to add onto a ggplot2 object.
Adapted from velocyto.R
}
