% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{winsorize_matrix}
\alias{winsorize_matrix}
\title{Internal function for winsorizing a matrix
Adapted from pagoda2: https://github.com/hms-dbmi/pagoda2}
\usage{
winsorize_matrix(mat, trim)
}
\description{
Internal function for winsorizing a matrix
Adapted from pagoda2: https://github.com/hms-dbmi/pagoda2
}
