% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{WriteGenesets}
\alias{WriteGenesets}
\title{Writes genesets to gmt file}
\usage{
WriteGenesets(genesets, file.name)
}
\arguments{
\item{genesets}{List of genesets}

\item{file.name}{Name of output file}
}
\value{
Writes to file
}
\description{
Writes genesets to gmt file
}
