% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{WriteGroups}
\alias{WriteGroups}
\title{Write sample groups from list to csv format}
\usage{
WriteGroups(groups.list, out.file)
}
\arguments{
\item{groups.list}{List of groups}

\item{out.file}{Name of output file}
}
\description{
Write sample groups from list to csv format
}
