% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{EmbedFeatures}
\alias{EmbedFeatures}
\title{Embeds features relative to factor coordinates}
\usage{
EmbedFeatures(swne.embedding, feature.assoc, features.embed, alpha.exp = 1,
  n_pull = NULL, scale.cols = T)
}
\arguments{
\item{swne.embedding}{Existing swne embedding from EmbedSWNE}

\item{feature.assoc}{Feature loadings or correlations (features x factors)}

\item{features.embed}{Names of features to embed}

\item{alpha.exp}{Increasing alpha.exp increases how much the factors "pull" the features}

\item{n_pull}{Number of factors pulling on each feature. Must be >= 3}

\item{scale.cols}{Whether or not to scale the input columns to 0 - 1}
}
\value{
swne.embedding with feature coordinates (feature.coords)
}
\description{
Embeds features relative to factor coordinates
}
