% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{EmbedSWNE}
\alias{EmbedSWNE}
\title{Embeds NMF factors and samples in 2 dimensions}
\usage{
EmbedSWNE(H, SNN = NULL, alpha.exp = 1, snn.exp = 1, n_pull = NULL,
  proj.method = "sammon", pca.red = F, dist.use = "cosine",
  snn.factor.proj = T, n.neighbors = max(round(nrow(H)/3), 4),
  min.dist = 0.5)
}
\arguments{
\item{H}{NMF factors (factors x samples)}

\item{SNN}{Shared nearest neighbors matrix (or other similarity matrix)}

\item{alpha.exp}{Increasing alpha.exp increases how much the NMF factors "pull" the samples}

\item{snn.exp}{Decreasing snn.exp increases the effect of the similarity matrix on the embedding}

\item{n_pull}{Number of factors pulling on each sample. Must be >= 3}

\item{proj.method}{Method to use for projecting factors. Currently only supports "sammon"}

\item{pca.red}{Whether or not to run PCA on transposed H matrix before calculating factor distances}

\item{dist.use}{Similarity function to use for calculating factor positions. Options include pearson (correlation), IC (mutual information), cosine, euclidean.}

\item{n.neighbors}{Number of neighbors for UMAP factor projection}

\item{min.dist}{Minimum distance for UMAP factor projection}
}
\value{
A list of factor (H.coords) and sample coordinates (sample.coords) in 2D
}
\description{
Embeds NMF factors and samples in 2 dimensions
}
