% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{EmbedSWNE}
\alias{EmbedSWNE}
\title{Projects NMF factors and samples in a 2D}
\usage{
EmbedSWNE(H, snn = NULL, alpha.exp = 1, snn.exp = 1, n_pull = NULL,
  proj.method = "sammon", dist.use = "cosine", min.snn = 0)
}
\arguments{
\item{H}{NMF factors (factors x samples)}

\item{snn}{Shared nearest neighbors matrix (or other similarity matrix)}

\item{alpha.exp}{Increasing alpha.exp increases how much the NMF factors "pull" the samples}

\item{snn.exp}{Decreasing snn.exp increases the effect of the similarity matrix on the embedding}

\item{n_pull}{Number of factors pulling on each sample. Must be >= 3}

\item{proj.method}{Method to use for projecting factors. Currently either "pca", or "sammon"}

\item{dist.use}{Similarity function to use for calculating factor positions. Options include pearson, IC, cosine.}

\item{min.snn}{Minimum SNN value}
}
\value{
A list of factor (H.coords) and sample coordinates (sample.coords) in 2D
}
\description{
Projects NMF factors and samples in a 2D
}
