% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{FeaturePlotDims}
\alias{FeaturePlotDims}
\title{Plots 2d embedding with feature overlayed}
\usage{
FeaturePlotDims(dim.scores, feature.scores, feature.name = NULL,
  x.lab = "tsne1", y.lab = "tsne2", alpha.plot = 0.5,
  quantiles = c(0.01, 0.99), show.axes = T, pt.size = 1,
  font.size = 12, color.palette = "YlOrRd")
}
\arguments{
\item{dim.scores}{2D embedding coordinates. Must be N x 2 samples}

\item{feature.name}{Name of feature}

\item{x.lab}{X axis label}

\item{y.lab}{Y axis label}

\item{alpha.plot}{Data point transparency}

\item{quantiles}{Quantiles to trim outliers from}

\item{show.axes}{Plot x and y axes}

\item{pt.size}{Sample point size}

\item{font.size}{Font size for axis labels}

\item{color.palette}{RColorbrewer palette to use}

\item{feature.score}{Feature vector to overlay}
}
\value{
ggplot2 object with dim plot with feature overlayed
}
\description{
Plots 2d embedding with feature overlayed
}
