% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{FeaturePlotSWNE}
\alias{FeaturePlotSWNE}
\title{Plots swne embedding with feature overlayed}
\usage{
FeaturePlotSWNE(swne.embedding, feature.scores, feature.name = NULL,
  alpha.plot = 0.5, quantiles = c(0.01, 0.99), samples.plot = NULL,
  label.size = 4.5, pt.size = 1)
}
\arguments{
\item{swne.embedding}{SWNE embedding (list of NMF and sample coordinates) from EmbedSWNE}

\item{feature.name}{Name of feature to overlay}

\item{alpha.plot}{Data point transparency}

\item{quantiles}{Quantiles to trim outliers from}

\item{samples.plot}{Samples to actually plot. Default is NULL, which plots all samples}

\item{label.size}{Label font size}

\item{pt.size}{Sample point size}

\item{feature.score}{Feature vector to overlay}
}
\value{
ggplot2 object with swne plot with feature overlayed
}
\description{
Plots swne embedding with feature overlayed
}
