% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{PlotSWNE}
\alias{PlotSWNE}
\title{Plots swne embedding}
\usage{
PlotSWNE(swne.embedding, alpha.plot = 0.25, sample.groups = NULL,
  do.label = F, label.size = 4.5, pt.size = 1, samples.plot = NULL,
  show.legend = T, seed = NULL)
}
\arguments{
\item{swne.embedding}{SWNE embedding (list of factor and sample coordinates) from EmbedSWNE}

\item{alpha.plot}{Data point transparency}

\item{sample.groups}{Factor defining sample groups}

\item{do.label}{Label the sample groups}

\item{label.size}{Label font size}

\item{pt.size}{Sample point size}

\item{samples.plot}{Vector of samples to plot. Default is NULL, which plots all samples.}

\item{show.legend}{If sample groups defined, show legend}

\item{seed}{Seed for sample groups color reproducibility}
}
\value{
ggplot2 object with swne plot
}
\description{
Plots swne embedding
}
