% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{ProjectGenesets}
\alias{ProjectGenesets}
\title{Projects dataset onto genesets using either NMF or a nonnegative linear model}
\usage{
ProjectGenesets(norm.counts, genesets, method = "lm", loss = "mse",
  n.cores = 1)
}
\arguments{
\item{norm.counts}{Normalized input data}

\item{genesets}{List of genesets to use}

\item{method}{Method used to project data onto genesets. lm uses a simple linear model while mf
will use a masked version of nmf.}

\item{loss}{Loss function for the nonnegative linear model}

\item{n.cores}{Number of cores to use}
}
\value{
List of gene loadings and factor scores, with each factor corresponding to a geneset
}
\description{
Projects dataset onto genesets using either NMF or a nonnegative linear model
}
