% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{ProjectSNN}
\alias{ProjectSNN}
\title{Calculates shared nearest neighbors between samples (columns) in a test matrix and samples in a training matrix
Adapted from Seurat}
\usage{
ProjectSNN(test.matrix, train.matrix, k = 20, k.scale = 10,
  prune.SNN = 1/15, print.output = T)
}
\arguments{
\item{test.matrix}{Test data}

\item{train.matrix}{Training data}

\item{k}{Number of nearest neighbors}

\item{k.scale}{k*k.scale is the number of nearest neighbors to calculate shared nearest neighbors for}

\item{prune.SNN}{Minimum fraction of shared nearest neighbors}

\item{print.output}{Prints progress bar if true}
}
\value{
A n.test x n.train matrix with the shared nearest neighbors between test and training data
}
\description{
Calculates shared nearest neighbors between samples (columns) in a test matrix and samples in a training matrix
Adapted from Seurat
}
