% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{ProjectSWNE}
\alias{ProjectSWNE}
\title{Projects new data onto an existing swne embedding}
\usage{
ProjectSWNE(swne.embedding, H.test, SNN.test = NULL, alpha.exp = 1,
  snn.exp = 1, n_pull = NULL)
}
\arguments{
\item{swne.embedding}{Existing swne embedding from EmbedSWNE}

\item{H.test}{Projected test data NMF scores}

\item{SNN.test}{Projected test data shared nearest neighbors matrix}

\item{alpha.exp}{Increasing alpha.exp increases how much the NMF components "pull" the samples}

\item{snn.exp}{Decreasing snn.exp increases the effect of the similarity matrix on the embedding}

\item{n_pull}{Number of NMF components pulling on each sample. n_pull >= 3}
}
\value{
Matrix of sample embeddings
}
\description{
Projects new data onto an existing swne embedding
}
