% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swne_plotting.R
\name{ProjectSWNE}
\alias{ProjectSWNE}
\title{Projects new data onto an existing swne embedding}
\usage{
ProjectSWNE(
  swne.embedding,
  H.test,
  SNN = NULL,
  alpha.exp = 1,
  snn.exp = 1,
  n_pull = NULL
)
}
\arguments{
\item{swne.embedding}{Existing swne embedding from EmbedSWNE}

\item{H.test}{Test factor scores}

\item{SNN}{Test SNN matrix}

\item{alpha.exp}{Increasing alpha.exp increases how much the factors "pull" the samples}

\item{snn.exp}{Decreasing snn.exp increases the effect of the similarity matrix on the embedding}

\item{n_pull}{Number of factors pulling on each sample. Must be >= 3}
}
\value{
SWNE embedding for the test data
}
\description{
Projects new data onto an existing swne embedding
}
