% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{RunNMF}
\alias{RunNMF}
\title{Runs NMF decomposition on a data matrix: A = WH.}
\usage{
RunNMF(A, k, alpha = 0, init = "ica", n.cores = 1, loss = "mse",
  n.rand.init = 3, init.zeros = "random")
}
\arguments{
\item{A}{Input data matrix}

\item{k}{Number of NMF factors}

\item{alpha}{Regularization parameter}

\item{init}{Initialization method: ica, nnsvd, or random.}

\item{n.cores}{Number of cores}

\item{loss}{Type of loss function to use}

\item{n.rand.init}{If random initialization is used, number of random restarts}

\item{init.zeros}{What to do with zeros in the initialization}
}
\value{
List of W (features x factors) and H (factors x samples)
}
\description{
Runs NMF decomposition on a data matrix: A = WH.
}
