% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swne.R
\name{RunSWNE}
\alias{RunSWNE}
\alias{RunSWNE.cisTopic}
\alias{RunSWNE.Seurat}
\alias{RunSWNE.Pagoda2}
\alias{RunSWNE.dgCMatrix}
\alias{RunSWNE.matrix}
\alias{RunSWNE.dgTMatrix}
\title{Wrapper for running SWNE analysis}
\usage{
\method{RunSWNE}{cisTopic}(cisTopicObject, proj.method = "sammon",
  cells.use = NULL, dist.metric = "cosine", n.cores = 8,
  hide.factors = T, n_pull = 3, alpha.exp = 1.25, snn.exp = 1,
  snn.k = 20, prune.SNN = 1/15)

\method{RunSWNE}{Seurat}(object, proj.method = "sammon",
  reduction.use = "pca", cells.use = NULL, dims.use = NULL,
  genes.use = NULL, dist.metric = "cosine", distance.matrix = NULL,
  n.cores = 8, k, k.range, var.genes, loss = "mse", genes.embed,
  hide.factors = T, n_pull = 3, ica.fast = T, alpha.exp = 1.25,
  snn.exp = 1, snn.k = 10, reduction.name = "swne",
  reduction.key = "SWNE_", return.format = c("embedding", "seurat"),
  ...)

\method{RunSWNE}{Pagoda2}(object, proj.method = "sammon",
  dist.metric = "cosine", n.cores = 8, k, k.range, var.genes,
  loss = "mse", genes.embed, hide.factors = T, n_pull = 3,
  ica.fast = T, n.var.genes = 3000, alpha.exp = 1.25, snn.exp = 1)

\method{RunSWNE}{dgCMatrix}(data.matrix, proj.method = "sammon",
  dist.metric = "cosine", n.cores = 3, k, k.range,
  var.genes = rownames(data.matrix), loss = "mse", genes.embed,
  hide.factors = T, n_pull = 3, ica.fast = T, alpha.exp = 1.25,
  snn.exp = 1)

\method{RunSWNE}{matrix}(data.matrix, proj.method = "sammon",
  dist.metric = "cosine", n.cores = 3, k, k.range,
  var.genes = rownames(data.matrix), loss = "mse", genes.embed,
  hide.factors = T, n_pull = 3, alpha.exp = 1.25, snn.exp = 1)

\method{RunSWNE}{dgTMatrix}(data.matrix, proj.method = "sammon",
  dist.metric = "cosine", n.cores = 3, k, k.range,
  var.genes = rownames(data.matrix), loss = "mse", genes.embed,
  hide.factors = T, n_pull = 3, alpha.exp = 1.25, snn.exp = 1)
}
\arguments{
\item{proj.method}{Method to use to project factors in 2D. Either "sammon" or "umap"}

\item{cells.use}{Which cells to analyze (default, all cells)}

\item{n.cores}{Number of cores to use (passed to FindNumFactors)}

\item{hide.factors}{Hide factors when plotting SWNE embedding}

\item{n_pull}{Maximum number of factors "pulling" on each sample}

\item{alpha.exp}{Increasing alpha.exp increases how much the NMF factors "pull" the samples (passed to EmbedSWNE)}

\item{snn.exp}{Decreasing snn.exp increases the effect of the similarity matrix on the embedding (passed to EmbedSWNE)}

\item{snn.k}{Changes the number of nearest neighbors used to build SNN (passed to CalcSNN)}

\item{object}{A Seurat, Pagoda2 or cisTopicObject object with normalized data}

\item{reduction.use}{Which dimensional reduction (e.g. PCA, ICA) to use for the tSNE. Default is PCA.}

\item{dims.use}{Which dimensions to use as input features}

\item{genes.use}{If set, run the SWNE on this subset of genes (instead of running on a set of reduced dimensions). Not set (NULL) by default}

\item{distance.matrix}{If set, runs tSNE on the given distance matrix instead of data matrix (experimental)}

\item{k}{Number of NMF factors (passed to RunNMF). If none given, will be derived from FindNumFactors.}

\item{k.range}{Range of factors for FindNumFactors to iterate over if k is not given}

\item{var.genes}{vector to specify variable genes. Will infer from Seurat or use full dataset if not given.}

\item{loss}{loss function to use (passed to RunNMF)}

\item{genes.embed}{Genes to add to the SWNE embedding}

\item{ica.fast}{Whether to run SVD before ICA initialization}

\item{reduction.name}{dimensional reduction name, specifies the position in the object$dr list. swne by default}

\item{reduction.key}{dimensional reduction key, specifies the string before the number for the dimension names. SWNE_ by default}

\item{return.format}{format to return ("seurat" object or raw "embedding")}

\item{n.var.genes}{Number of variable genes to use}

\item{data.matrix}{a data matrix (genes x cells) which has been pre-normalized}

\item{batch}{Vector of batch IDs to regress away}

\item{dist.use}{Similarity function to use for calculating factor positions (passed to EmbedSWNE).
Options include pearson (correlation), IC (mutual information), cosine, euclidean.}
}
\value{
A list of factor (H.coords) and sample coordinates (sample.coords) in 2D
}
\description{
Wrapper for running SWNE analysis
}
