% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmf.R
\name{SummarizeAssocFeatures}
\alias{SummarizeAssocFeatures}
\title{Summarize the factors x features association matrix into a readable dataframe}
\usage{
SummarizeAssocFeatures(
  feature.factor.assoc,
  features.return = 10,
  features.use = NULL
)
}
\arguments{
\item{feature.factor.assoc}{(features x factors) association matrix from FactorAssociation}

\item{features.return}{Number of top features to return for each factor}

\item{features.use}{Only consider a subset of features. Default is NULL, which will use all features}
}
\value{
Dataframe summarizing top features for each factor
}
\description{
Summarize the factors x features association matrix into a readable dataframe
}
